% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{add_datapoints}
\alias{add_datapoints}
\title{Add Age Control Points to Plot}
\usage{
add_datapoints(gg, dat)
}
\arguments{
\item{gg}{a ggplot2 object}

\item{dat}{a dataframe containing the datapoints to add}
}
\value{
modifies a ggplot2 object
}
\description{
Add Age Control Points to Plot
}
\keyword{internal}
