% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{add_subdivisions}
\alias{add_subdivisions}
\title{Title}
\usage{
add_subdivisions(gg, hamstr_fit)
}
\arguments{
\item{gg}{}

\item{hamstr_fit}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
