% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{AdjustK}
\alias{AdjustK}
\title{Adjust numbers of splits per level}
\usage{
AdjustK(K_fine, base)
}
\arguments{
\item{K_fine}{total number of required sections at highest resolution}
}
\value{

}
\description{
Adjust numbers of splits per level
}
\keyword{internal}
