% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{alpha_indices}
\alias{alpha_indices}
\title{Create alpha level indices}
\usage{
alpha_indices(K)
}
\arguments{
\item{K}{controls the number and structure of the hierarchically modelled
down-core sections. It is specified as a vector, where each value indicates
the number of new child sections for each parent section, e.g., c(10, 10,
10) would specify 10 sections at the coarsest level, with 10 new sections
at next finer level, giving a total of 1000 sections at the finest
resolution.

By default, the total number of sections at the finest level is set so that
for the median distance between age control points there are 16 sections.

By default a "powers of 2" hierarchical structure is chosen and then
adjusted to get close to the desired number of sections, e.g. for a desired
100 sections the result is c(2, 2, 2, 2, 2, 3), giving 96 sections.}
}
\value{
a list
}
\description{
Create alpha level indices
}
\keyword{internal}
