% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baconr-package.R
\docType{package}
\name{baconr-package}
\alias{baconr-package}
\alias{baconr}
\title{baconr}
\description{
baconr
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
