% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiocarbon-functions.R
\name{compare_14C_PDF}
\alias{compare_14C_PDF}
\title{Compare the full empirical calendar age PDF of a radiocarbon date with a
t-distribution approximation}
\usage{
compare_14C_PDF(
  age.14C,
  age.14C.se,
  cal_curve = "intcal13",
  t_df = 6,
  return.type = c("plot", "list")
)
}
\arguments{
\item{age.14C}{vector of radiocarbon dates in years BP}

\item{age.14C.se}{vector of radiocarbon date uncertainties}

\item{cal_curve}{calibration curve}

\item{t_df}{degrees of freedom of the t-distribution}

\item{return.type}{return a ggplot object or a list containing the ggplot
object and two data frames with the empirical and t-distributions}
}
\value{

}
\description{
Compare the full empirical calendar age PDF of a radiocarbon date with a
t-distribution approximation
}
\examples{
compare_14C_PDF(age.14C = c(1000, 4000), age.14C.se = c(100, 150),
 cal_curve = "intcal13", return.type = "plot")
}
