% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{default_K}
\alias{default_K}
\title{Default K structure}
\usage{
default_K(K_fine)
}
\arguments{
\item{K_fine}{total number of sections at the finest resolution}
}
\value{
a vector
}
\description{
Default K structure
}
\examples{
\dontrun{
default_K(100)
default_K(500)  
}
}
\keyword{internal}
