% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{extract_hamstr_fit}
\alias{extract_hamstr_fit}
\title{Extract posterior age models or parameters}
\usage{
extract_hamstr_fit(object, pars = c("ages"), ...)
}
\arguments{
\item{object}{hamstr_fit object}

\item{pars}{The parameters to extract. If pars = "ages" (the default) a special
method is invoke to return the age-depth models. If pars = "pars" the memory parameters
and the overall mean accumulation rate are returned. Any other specification
invokes the default rstan as.data.frame extract method for the use specified parameters.}

\item{...}{Other arguments to rstan::extract}
}
\value{

}
\description{
Extract posterior age models or parameters
}
