% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{filter_hamstr_acc_rates}
\alias{filter_hamstr_acc_rates}
\title{Smooth accumulation rates}
\usage{
filter_hamstr_acc_rates(hamstr_acc_rates, tau, kern)
}
\arguments{
\item{hamstr_acc_rates}{Accumulation rates from get_posterior_acc_rates}

\item{tau}{Scale of the smoothing kernel in depth units. If tau > 0,
accumulation rates are smoothed (filtered) before summary statistics are 
calculated, so that the accumulation rate at a given depth corresponds to the
average rate over the depth interval tau. Default to 0.}

\item{kern}{Choice of smoothin kernal. U for uniform (moving average), G for
Gaussian, BH for Berger and Heath (exponential). Defaults to U}
}
\value{

}
\description{
Smooth accumulation rates
}
\keyword{internal}
