% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{gamma_sigma_shape}
\alias{gamma_sigma_shape}
\title{Convert between parametrisations of the gamma distribution}
\usage{
gamma_sigma_shape(mean = NULL, mode = NULL, sigma = NULL, shape = NULL)
}
\arguments{
\item{mean}{Mean of gamma distribution}

\item{mode}{Mode of the gamma distribution}

\item{sigma}{Standard deviation of gamma distribution}

\item{shape}{Shape of gamma distribution}
}
\value{
a list
}
\description{
Convert between parametrisations of the gamma distribution
}
\examples{
\dontrun{
gamma_sigma_shape(mean = 10, sigma = 2)
}
}
\keyword{internal}
