% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_adam.R
\name{get_inits_adam}
\alias{get_inits_adam}
\title{Create Random Initial Values for the adam Stan Model}
\usage{
get_inits_adam(stan_dat)
}
\arguments{
\item{stan_dat}{stan data for the adam model}
}
\value{
a list of lists
}
\description{
Create Random Initial Values for the adam Stan Model
}
\keyword{internal}
