% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{get_inits_hamstr}
\alias{get_inits_hamstr}
\title{Create Random Initial Values for the hamstr Stan Model}
\usage{
get_inits_hamstr(stan_dat)
}
\arguments{
\item{stan_dat}{stan data for the hamstr model}
}
\value{
a list of lists
}
\description{
Create Random Initial Values for the hamstr Stan Model
}
\keyword{internal}
