% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_par_idx}
\alias{get_par_idx}
\title{Extract Stan Parameter Index}
\usage{
get_par_idx(x)
}
\arguments{
\item{x}{vector of stan parameter names}
}
\value{

}
\description{
Extract Stan Parameter Index
}
\keyword{internal}
