% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{get_posterior_parameters}
\alias{get_posterior_parameters}
\title{Get Posterior Parameters}
\usage{
get_posterior_parameters(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{
a dataframe/tibble with posterior ages for all iterations after warmup
}
\description{
Get Posterior Parameters
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error)

get_posterior_parameters(fit)
}
}
