% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{get_smooth_i}
\alias{get_smooth_i}
\title{Get Indices For Smoothing Accumulation Rates When Estimating L}
\usage{
get_smooth_i(d, w)
}
\arguments{
\item{d}{}

\item{w}{}
}
\description{
Get Indices For Smoothing Accumulation Rates When Estimating L
}
\keyword{internal}
