% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{GetK}
\alias{GetK}
\title{Create K structure from K_tot and target_K_per_lvl}
\usage{
GetK(K_tot, target_K_per_lvl = 10)
}
\arguments{
\item{K_tot}{total number of required sections at highest resolution}

\item{target_K_per_lvl}{approximate number of sections per level}
}
\value{

}
\description{
Create K structure from K_tot and target_K_per_lvl
}
\keyword{internal}
