% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr_prime.R
\name{GetNextMultiPrimes}
\alias{GetNextMultiPrimes}
\title{Get Nearest Primes to Geometric Series}
\usage{
GetNextMultiPrimes(m, f = 2)
}
\arguments{
\item{m}{Terminal number}

\item{f}{Approximate ratio of series}
}
\description{
Get Nearest Primes to Geometric Series
}
\examples{
GetNextMultiPrimes(100, 3)
}
