% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{GetWts}
\alias{GetWts}
\title{Get Weights for Parent Sections}
\usage{
GetWts(a, b)
}
\arguments{
\item{a}{Parent breaks}

\item{b}{Child breaks}
}
\description{
Get Weights for Parent Sections
}
\keyword{internal}
