% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_hamstr.R
\name{hamstr_control}
\alias{hamstr_control}
\title{Additional Arguments to hamstr useful for debugging or development}
\usage{
hamstr_control(
  scale_R = TRUE,
  nu = 6,
  scale_shape = TRUE,
  smooth_s = FALSE,
  inflate_errors = FALSE,
  infl_sigma_sd = NULL,
  infl_shape_shape = 1,
  infl_shape_mean = 1
)
}
\arguments{
\item{scale_R}{logical: Scale AR1 coefficient by delta_c (as in Bacon) or
not. Defaults to TRUE.}

\item{nu}{Degrees of freedom for the Student-t distributed error model.
Defaults to 6, which is equivalent to the default parameterisation of
t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian
error model, (nu = 100 should do it).}

\item{scale_shape}{Scale the shape parameter according to the number of
hierarchical levels, to control the total variance of the alpha
innovations. This defaults to TRUE as of Hamstr verion 0.5.}

\item{smooth_s}{Smooth the sedimentation rate used to calculate additional
error from bioturbation by taking a mean across nearby sections}

\item{inflate_errors}{logical: If set to TRUE, observation errors are
inflated so that data are consistent with a gamma AR1 age-depth model. 
This is an experimental feature under active development. Defaults to FALSE.}

\item{infl_sigma_sd}{Hyperparameter: sets the standard deviation of the
half-normal prior on the mean of the additional error terms. Defaults to 10
times the mean observation error in obs_err.}

\item{infl_shape_shape, infl_shape_mean}{Hyperparameters: parametrises the
gamma prior on the shape of the distribution of the additional error terms.
Default to 1, 1.}
}
\value{

}
\description{
Additional Arguments to hamstr useful for debugging or development
}
