% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{hierarchical_depths}
\alias{hierarchical_depths}
\title{Calculated depth of section boundary at all hierarchical levels}
\usage{
hierarchical_depths(stan_dat)
}
\arguments{
\item{stan_dat}{}
}
\value{

}
\description{
Calculated depth of section boundary at all hierarchical levels
}
\keyword{internal}
