% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{hierarchy_efficiency}
\alias{hierarchy_efficiency}
\title{Calculate number of parameters being estimated for a given hierarchical structure}
\usage{
hierarchy_efficiency(base, n)
}
\arguments{
\item{base}{Number of new sections per per section}

\item{n}{Number of hierarchical levels}
}
\value{

}
\description{
Calculate number of parameters being estimated for a given hierarchical structure
}
\examples{
hierarchy_efficiency(10, 3)
}
