% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat.R
\name{make_stan_dat}
\alias{make_stan_dat}
\title{make_stan_dat}
\usage{
make_stan_dat(depth, obs_age, obs_err, K = 10, nu = 6, acc_mean,
  acc_var = "default", mem_mean = 0.7, mem_strength = 4)
}
\arguments{
\item{depth}{Vector of depths from and observed age-depth profile}

\item{obs_age}{Observed age at each depth}

\item{obs_err}{Error associated with each observed age (1 standard error)}

\item{K}{Number of sections into which the profile will be divided}

\item{nu}{Degrees of freedom for the Student-t distributed error model.
Defaults to 6, which is equivalent to the default parameterisation of
t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian
error model, (nu = 100 should do it).}

\item{acc_mean}{Hyper-parameter: the mean sediment accumulation rate for the
Gamma prior on sedimentation rate innovations}

\item{acc_var}{Hyper-parameter: the variance of the sediment accumulation
rate for the Gamma prior on sedimentation rate innovations. If set to
"default" acc_var = acc_mean^2 / 1.5, which is equivalent to the default
acc.shape = 1.5 parameter in Bacon 2.2}

\item{mem_mean}{Hyper-parameter: a parameter of the Beta prior distribution
on "memory", i.e. the autocorrelation parameter in the underlying AR1
model. The prior on the correlation between layers is scaled according to the
thickness of the layers \emph{delta_c}, which is determined by the total length of the
profile and the parameter \emph{K}. mem_mean sets the mean value for \emph{R}
(defaults to 0.7), while \emph{w} = R^(delta_c/1)}

\item{mem_strength}{Hyper-parameter: sets the strength of the memory prior,
defaults to 4 as in Bacon 2.2}
}
\description{
make_stan_dat
}
