% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr_prime.R
\name{make_stan_dat_hamstr_prime}
\alias{make_stan_dat_hamstr_prime}
\title{Make the data object required by the Stan sampler}
\usage{
make_stan_dat_hamstr_prime(...)
}
\arguments{
\item{...}{Arguments passed from \code{\link{hamstr}}}
}
\value{
a list of data and parameters to be passed as data to the Stan sampler
}
\description{
Make the data object required by the Stan sampler
}
\keyword{internal}
