% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_sarnie.R
\name{make_stan_dat_sarnie}
\alias{make_stan_dat_sarnie}
\title{make_stan_dat}
\usage{
make_stan_dat_sarnie(
  depth,
  obs_age,
  obs_err,
  K1 = 10,
  K = 100,
  nu = 6,
  record_prior_acc_mean_mean = 20,
  record_prior_acc_mean_shape = 1.5,
  record_prior_acc_shape_mean = 1.5,
  record_prior_acc_shape_shape = 1.5,
  mem_mean = 0.7,
  mem_strength = 4,
  inflate_errors = 0
)
}
\arguments{
\item{depth}{Vector of depths from an observed age-depth profile}

\item{obs_age}{Observed age at each depth}

\item{obs_err}{Error associated with each observed age (1 standard error)}

\item{K}{Number of sections into which the profile will be divided}

\item{nu}{Degrees of freedom for the Student-t distributed error model.
Defaults to 6, which is equivalent to the default parameterisation of
t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian
error model, (nu = 100 should do it).}

\item{mem_mean}{Hyper-parameter: a parameter of the Beta prior distribution
on "memory", i.e. the autocorrelation parameter in the underlying AR1
model. The prior on the correlation between layers is scaled according to
the thickness of the layers *delta_c*, which is determined by the total
length of the profile and the parameter *K*. mem_mean sets the mean value
for *R* (defaults to 0.7), while *w* = R^(delta_c)}

\item{mem_strength}{Hyper-parameter: sets the strength of the memory prior,
defaults to 4 as in Bacon 2.2}
}
\value{
a list of data, and parameters to be passed as data to the Stan sampler
}
\description{
make_stan_dat
}
\examples{
make_stan_dat_sarnie(depth = MSB2K$depth,
              obs_age = MSB2K$age,
              obs_err = MSB2K$error,
              K1 = 10, K = 100,
              nu = 6)
}
