% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_adam.R
\name{optimal_K}
\alias{optimal_K}
\title{Estimate Optimal Hierarchical Structure}
\usage{
optimal_K(K_tot, target_K_per_lvl = 10)
}
\arguments{
\item{K_tot}{}
}
\value{

}
\description{
Estimate Optimal Hierarchical Structure
}
\keyword{internal}
