% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{OptK}
\alias{OptK}
\title{Find a good hierarchical structure}
\usage{
OptK(K_fine)
}
\arguments{
\item{K_fine}{}
}
\value{

}
\description{
Find a good hierarchical structure
}
\keyword{internal}
