% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot.hamstr_fit}
\alias{plot.hamstr_fit}
\title{Title}
\usage{
\method{plot}{hamstr_fit}(
  object,
  type = c("default", "age_models", "hier_acc_rates", "acc_mean_prior_post",
    "mem_prior_post"),
  summarise = TRUE,
  ...
)
}
\arguments{
\item{object}{a hamstr_fit object}

\item{type}{One of "default", "age_models", "hier_acc_rates",
"acc_mean_prior_post", "mem_prior_post"}

\item{summarise}{}
}
\value{
A ggplot object
}
\description{
Title
}
