% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_acc_mean_prior_posterior}
\alias{plot_acc_mean_prior_posterior}
\title{Plot Mean Accumulation Rate Prior and Posterior Distributions}
\usage{
plot_acc_mean_prior_posterior(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{}
}
\value{

}
\description{
Plot Mean Accumulation Rate Prior and Posterior Distributions
}
