% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_bacon.R
\name{plot_adam}
\alias{plot_adam}
\title{Plot an adam_fit object}
\usage{
plot_adam(
  adam_fit,
  type = c("ribbon", "spaghetti"),
  n.iter = 1000,
  plot_diagnostics = FALSE
)
}
\arguments{
\item{adam_fit}{The object returned from \code{stan_bacon}.}

\item{n.iter}{The number of iterations of the model to plot, defaults to 1000.}
}
\value{
A ggplot2 object
}
\description{
Plots the Bacon modelled Age~Depth relationship together with
  the depths, ages, and age uncertainties in the observed data. A sample of
  size \code{n.iter} of the interactions of the posterior distribution are
  plotted as grey lines. The observed data are plotted as points with +- 2*se error bars.
}
