% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_age_models}
\alias{plot_age_models}
\title{Plot Age Models as Spaghetti Plot}
\usage{
plot_age_models(hamstr_fit, n.iter = 1000)
}
\arguments{
\item{hamstr_fit}{the object returned from \code{stan_hamstr}.}

\item{n.iter}{the number of iterations of the model to plot, defaults to
1000.}
}
\value{
A ggplot2 object
}
\description{
Plot Age Models as Spaghetti Plot
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error)

plot_age_models(fit)
}
}
\keyword{internal}
