% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_D_prior_posterior}
\alias{plot_D_prior_posterior}
\title{Plot Displacement Depth Prior and Posterior}
\usage{
plot_D_prior_posterior(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{
A ggplot2 object
}
\description{
Plot Displacement Depth Prior and Posterior
}
\keyword{internal}
