% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_downcore_summary}
\alias{plot_downcore_summary}
\title{Plot Downcore Summary}
\usage{
plot_downcore_summary(ds)
}
\arguments{
\item{ds}{a downcore summary of age or accumulation rate}
}
\value{

}
\description{
Plot Downcore Summary
}
\keyword{internal}
