% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_gamma_prior_posterior}
\alias{plot_gamma_prior_posterior}
\title{Plot A Gamma Prior and Posterior}
\usage{
plot_gamma_prior_posterior(hamstr_fit, par)
}
\arguments{
\item{hamstr_fit}{the object returned from \code{stan_hamstr}.}

\item{par}{parameter to plot}
}
\value{
A ggplot2 object
}
\description{
Plot A Gamma Prior and Posterior
}
\keyword{internal}
