% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_hamstr}
\alias{plot_hamstr}
\title{Plot an hamstr_fit object}
\usage{
plot_hamstr(
  hamstr_fit,
  type = c("ribbon", "spaghetti"),
  n.iter = 1000,
  plot_diagnostics = TRUE
)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}

\item{n.iter}{The number of iterations of the model to plot, defaults to 1000.}
}
\value{
A ggplot2 object
}
\description{
Plots the Bacon modelled Age~Depth relationship together with
  the depths, ages, and age uncertainties in the observed data. A sample of
  size \code{n.iter} of the interactions of the posterior distribution are
  plotted as grey lines. The observed data are plotted as points with +- 2*se error bars.
}
