% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_hierarchical_acc_rate}
\alias{plot_hierarchical_acc_rate}
\title{Plot the hierarchical accumulation rate parameters}
\usage{
plot_hierarchical_acc_rate(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{}
}
\value{

}
\description{
Plot the hierarchical accumulation rate parameters
}
