% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_infl_prior_posterior}
\alias{plot_infl_prior_posterior}
\title{Plot the Prior and Posterior Distributions of the Inflation Factor Parameters}
\usage{
plot_infl_prior_posterior(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{
A ggplot2 object
}
\description{
Plot the Prior and Posterior Distributions of the Inflation Factor Parameters
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error,   
  inflate_errors = 1)

plot_infl_prior_posterior(fit)
}
}
\keyword{internal}
