% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_memory_prior_posterior}
\alias{plot_memory_prior_posterior}
\title{Plot Memory Prior and Posterior}
\usage{
plot_memory_prior_posterior(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{the object returned from \code{stan_hamstr}.}
}
\value{
a ggplot2 object
}
\description{
Plot Memory Prior and Posterior
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error)

plot_memory_prior_posterior(fit)
}
}
\keyword{internal}
