% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{plot_summary_age_models}
\alias{plot_summary_age_models}
\title{Plot Summary of Posterior Age Models}
\usage{
plot_summary_age_models(adam_fit)
}
\arguments{
\item{adam_fit}{}
}
\value{

}
\description{
Plot Summary of Posterior Age Models
}
