% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_hamstr_acc_rates}
\alias{summarise_hamstr_acc_rates}
\title{Summarise accumulation rates}
\usage{
summarise_hamstr_acc_rates(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{}
}
\value{

}
\description{
Summarise accumulation rates
}
\keyword{internal}
