% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_hamstr_parameters}
\alias{summarise_hamstr_parameters}
\title{Summarise Hamstr Parameters}
\usage{
summarise_hamstr_parameters(
  object,
  pars = c("alpha[1]", "R", "w", "L", "D", "H_depth", "H_length")
)
}
\arguments{
\item{object}{A hamstr_fit object}

\item{par}{Character vector of parameters to include}
}
\value{

}
\description{
Summarise Hamstr Parameters
}
\keyword{internal}
