% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_new_ages}
\alias{summarise_new_ages}
\title{Summarise Interpolated Posterior Age Models}
\usage{
summarise_new_ages(
  new_ages,
  probs = c(0.025, 0.159, 0.25, 0.5, 0.75, 0.841, 0.975)
)
}
\arguments{
\item{new_ages}{Object of class "hamstr_interpolated_ages"}
}
\value{
data.frame / tibble
}
\description{
Summarise Interpolated Posterior Age Models
}
\keyword{internal}
