% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summary.hamstr_fit}
\alias{summary.hamstr_fit}
\title{Title}
\usage{
\method{summary}{hamstr_fit}(object, type = "age_models")
}
\arguments{
\item{object}{}

\item{type}{}
}
\value{

}
\description{
Title
}
