% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summary.hamstr_interpolated_ages}
\alias{summary.hamstr_interpolated_ages}
\title{Title}
\usage{
\method{summary}{hamstr_interpolated_ages}(object, probs)
}
\arguments{
\item{object}{hamstr_interpolated_ages object}
}
\value{

}
\description{
Title
}
