% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{add_subdivisions}
\alias{add_subdivisions}
\title{Add subdivision tickmarks}
\usage{
add_subdivisions(gg, hamstr_fit)
}
\arguments{
\item{gg}{}

\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{
A ggplot2 object
}
\description{
Add subdivision tickmarks
}
\keyword{internal}
