% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_adam.R
\name{alpha_indices}
\alias{alpha_indices}
\title{Create alpha level indices}
\usage{
alpha_indices(K)
}
\arguments{
\item{K}{a vector with the number of new sub-divisions at each level,
defaults to c(10, 10), resulting in two levels with 10 and 100 sections
respectively.}
}
\value{
a list
}
\description{
Create alpha level indices
}
\keyword{internal}
