% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baconr-package.R
\docType{package}
\name{baconr-package}
\alias{baconr-package}
\alias{baconr}
\title{The 'baconr' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
