% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiocarbon-functions.R
\name{compare_14C_PDF}
\alias{compare_14C_PDF}
\title{Compare Empirical and t-distribution Approximated Calendar Age PDFs}
\usage{
compare_14C_PDF(
  age.14C,
  age.14C.se,
  offset = 0,
  offset.se = 0,
  cal_curve = "intcal20",
  nu = 6,
  return.type = c("plot", "list")
)
}
\arguments{
\item{age.14C}{vector of radiocarbon dates in years BP}

\item{age.14C.se}{vector of radiocarbon date uncertainties}

\item{offset}{vector of offsets, e.g. reservoir ages.}

\item{offset.se}{vector of offset uncertainties, e.g. sigmaDelatR.}

\item{cal_curve}{calibration curve}

\item{nu}{degrees of freedom of the t-distribution approximation, default in
hamstr is 6}

\item{return.type}{return a ggplot object or a list containing the ggplot
object and two data frames with the empirical and t-distributions}
}
\value{
A ggplot2 object or list with data and ggplot2 object
}
\description{
Compare the full empirical calendar age PDFs of calibrated
radiocarbon dates with the t-distribution approximations use by hamstr
}
\examples{
compare_14C_PDF(age.14C = c(1000, 4000), age.14C.se = c(100, 150),
 cal_curve = "intcal20", return.type = "plot")
}
