% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{get_K_factor}
\alias{get_K_factor}
\title{Get Default K_factor}
\usage{
get_K_factor(K_fine)
}
\arguments{
\item{K_fine}{The number of sections at the highest resolution}
}
\description{
Get Default K_factor
}
\examples{

K <- seq(1, 1e05, length.out = 1000)
K_factor <- sapply(K, function(x) get_K_factor(x))
plot(K, K_factor, type = "l", log = "xy")

}
\keyword{internal}
