% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{get_K_factor}
\alias{get_K_factor}
\title{Get Default K_factor}
\usage{
get_K_factor(K_fine)
}
\arguments{
\item{K_fine}{The number of sections at the highest resolution}
}
\description{
Get Default K_factor
}
\keyword{internal}
