% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{get_posterior_ages}
\alias{get_posterior_ages}
\title{Get Posterior Age Models}
\usage{
get_posterior_ages(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{fitted hamstr model, output from hamstr()}
}
\value{
a dataframe/tibble with posterior ages for all iterations after warmup
}
\description{
Get Posterior Age Models
}
\examples{
\dontrun{
get_posterior_ages(fitted.hamstr.object)
}
}
