% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{get_posterior_ages}
\alias{get_posterior_ages}
\title{Get Posterior Age Models}
\usage{
get_posterior_ages(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{the object returned from \code{stan_hamstr}.}
}
\value{
a dataframe/tibble with posterior ages for all iterations after warmup
}
\description{
Get Posterior Age Models
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error)

get_posterior_ages(fit)
}
}
\keyword{internal}
