% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{GetBrksHalfOffset}
\alias{GetBrksHalfOffset}
\title{Get Overlapping Breaks Structure}
\usage{
GetBrksHalfOffset(K_fine, K_factor)
}
\arguments{
\item{K_fine}{}

\item{K_factor}{}
}
\description{
Get Overlapping Breaks Structure
}
\keyword{internal}
