% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{GetIndices}
\alias{GetIndices}
\title{Get Indices Structure For Hamstr Model}
\usage{
GetIndices(nK = NULL, brks = NULL)
}
\arguments{
\item{nK}{}

\item{brks}{}
}
\description{
Get Indices Structure For Hamstr Model
}
\keyword{internal}
