% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{interpolate_age_models}
\alias{interpolate_age_models}
\title{Interpolate Posterior Age Model At New Depths}
\usage{
interpolate_age_models(hamstr_fit, new_depth = NULL)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}

\item{new_depth}{a vector of depths at which to interpolate the age models. 
If left NULL, the depths of the age control points are used.}
}
\value{
hamstr_interpolated_ages object
}
\description{
Interpolate Posterior Age Model At New Depths
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error,
  K = c(10, 10), nu = 6,
  acc_mean_prior = 20,
  mem_mean = 0.5, mem_strength = 10,
  inflate_errors = 0,
  iter = 2000, chains = 3)
  
interpolate.age.models(fit, new_depth = seq(1000, 15000, by = 1000))
}

}
\keyword{internal}
