% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{interpolate_age_models}
\alias{interpolate_age_models}
\title{Interpolate Posterior Age Model At New Depths}
\usage{
interpolate_age_models(hamstr_fit, new_depth)
}
\arguments{
\item{hamstr_fit}{fitted hamstr model, output from hamstr()}

\item{new_depth}{a vector of depths at which to interpolate the age models}
}
\value{

}
\description{
Interpolate Posterior Age Model At New Depths
}
\examples{
\dontrun{
interpolate.age.models(fitted.hamstr.object, new_depth = seq(1000, 15000, by = 1000))
}

}
