% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{make_stan_dat_hamstr}
\alias{make_stan_dat_hamstr}
\title{Make the data object required by the Stan program}
\usage{
make_stan_dat_hamstr(...)
}
\arguments{
\item{...}{Arguments passed from \code{\link{hamstr}}}
}
\value{
a list of data and parameters to be passed as data to the Stan sampler
}
\description{
Make the data object required by the Stan program
}
\keyword{internal}
