% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_dat_hamstr.R
\name{make_stan_dat_hamstr}
\alias{make_stan_dat_hamstr}
\title{Make the data object required by the Stan program}
\usage{
make_stan_dat_hamstr(
  depth = NULL,
  obs_age = NULL,
  obs_err = NULL,
  top_depth = NULL,
  bottom_depth = NULL,
  pad_top_bottom = NULL,
  K = NULL,
  nu = NULL,
  acc_mean_prior = NULL,
  acc_shape = NULL,
  scale_shape = NULL,
  mem_mean = NULL,
  mem_strength = NULL,
  scale_R = NULL,
  inflate_errors = NULL,
  infl_sigma_sd = NULL,
  infl_shape_shape = NULL,
  infl_shape_mean = NULL,
  ...
)
}
\arguments{
\item{depth}{Depths of observed ages (age control points)}

\item{obs_age}{Observed age at each depth (age control points)}

\item{obs_err}{Error associated with each observed age (1 standard error)}

\item{top_depth}{The top and bottom depths of the desired
age-depth model. Must encompass the range of the data. Defaults to the
shallowest and deepest data points unless \code{pad_top_bottom = TRUE}}

\item{bottom_depth}{The top and bottom depths of the desired
age-depth model. Must encompass the range of the data. Defaults to the
shallowest and deepest data points unless \code{pad_top_bottom = TRUE}}

\item{pad_top_bottom}{logical, pad the length of the age-depth model by 5% on
each end}

\item{K}{A vector with the number of new sub-divisions at each hierarchical
level, defaults to c(10, 10), resulting in two levels with 10 and 100
sections respectively.}

\item{nu}{Degrees of freedom for the Student-t distributed error model.
Defaults to 6, which is equivalent to the default parameterisation of
t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian
error model, (nu = 100 should do it).}

\item{acc_mean_prior}{Hyperparameter for the prior on the overall mean
accumulation rate for the record. Units are obs_age / depth. E.g. if depth
is in cm and age in kyr then the accumulation rate is in kyr/cm. The
overall mean accumulation rate is given a weak half-normal prior with mean
= 0, SD = 10 * acc_mean_prior. If left blank, acc_mean_prior is set to the
mean accumulation rate estimated by fitting a robust linear model using
\link[MASS]{rlm}.}

\item{acc_shape}{Hyperparameter for the shape of the priors on accumulation rates.
Defaults to 1.5 - as for Bacon 2.2.}

\item{scale_shape}{Scale the shape parameter according to the number of hierarchical
levels, to control the total variance of the alpha innovations. This defaults
to TRUE as of Hamstr verion 0.5.}

\item{mem_mean}{Hyperparameter: a parameter of the Beta prior distribution on
"memory", i.e. the autocorrelation parameter in the underlying AR1 model.
The prior on the correlation between layers is scaled according to the
thickness of the sediment sections in the highest resolution hierarchical
layer, *delta_c*, which is determined by the total length age-models and
the parameter vector *K*. mem_mean sets the mean value for *R* (defaults to
0.5), while *w* = R^(delta_c)}

\item{mem_strength}{Hyperparameter: sets the strength of the memory prior,
defaults to 10 as in Bacon >= 2.5.1}

\item{scale_R}{logical: Scale AR1 coefficient by delta_c (as in Bacon) or
not. Defaults to TRUE.}

\item{inflate_errors}{logical: If set to TRUE, observation errors are
inflated so that data are consistent with a "Bacon-style" monotonic
age-depth model. This is an experimental feature under active development.
Defaults to FALSE.}

\item{infl_sigma_sd}{Hyperparameter: sets the standard deviation of the
half-normal prior on the mean of the additional error terms. Defaults to 10
times the mean observation error in obs_err.}

\item{infl_shape_shape}{Hyperparameters: parametrises the
gamma prior on the shape of the distribution of the additional error terms.
 Default to 1, 1.}

\item{infl_shape_mean}{Hyperparameters: parametrises the
gamma prior on the shape of the distribution of the additional error terms.
 Default to 1, 1.}

\item{...}{additional arguments to \link[rstan]{sampling}}
}
\value{
a list of data and parameters to be passed as data to the Stan sampler
}
\description{
Make the data object required by the Stan program
}
\examples{
make_stan_dat_hamstr(depth = MSB2K$depth,
              obs_age = MSB2K$age,
              obs_err = MSB2K$error,
              K = 128)
}
