% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot.hamstr_fit}
\alias{plot.hamstr_fit}
\title{Title}
\usage{
\method{plot}{hamstr_fit}(
  x,
  type = c("default", "age_models", "acc_rates", "hier_acc_rates",
    "acc_mean_prior_post", "mem_prior_post", "L_prior_post", "D_prior_post", "PDF_14C"),
  summarise = TRUE,
  ...
)
}
\arguments{
\item{x}{a hamstr_fit object}

\item{type}{one of "default", "age_models","acc_rates", "hier_acc_rates",
"acc_mean_prior_post", "mem_prior_post", "L_prior_post", "D_prior_post", "PDF_14C"}

\item{summarise}{logical TRUE or FALSE. Plot the realisations as a summarised
"ribbon" showing 50% and 95% intervals (faster), or as a spaghetti plot
showing individual realisations. Defaults to TRUE (ribbon).}

\item{...}{additional arguments to hamstr plotting methods}
}
\value{
A ggplot object
}
\description{
Title
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error)

plot(fit)
plot(fit, type = "acc_rates", tau = 5, kern = "U")
}
}
