% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot.hamstr_fit}
\alias{plot.hamstr_fit}
\title{Title}
\usage{
\method{plot}{hamstr_fit}(
  object,
  type = c("default", "age_models", "acc_rates", "hier_acc_rates",
    "acc_mean_prior_post", "mem_prior_post"),
  summarise = TRUE,
  ...
)
}
\arguments{
\item{object}{a hamstr_fit object}

\item{type}{One of "default", "age_models", "hier_acc_rates",
"acc_mean_prior_post", "mem_prior_post"}

\item{summarise}{logical TRUE or FALSE. Plot the realisations as a summarised
"ribbon" showing 50% and 95% intervals (faster), or as a spaghetti plot 
showing individual realisations. Defaults to TRUE (ribbon).}
}
\value{
A ggplot object
}
\description{
Title
}
