% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_hamstr}
\alias{plot_hamstr}
\title{Plot an hamstr_fit object}
\usage{
plot_hamstr(
  hamstr_fit,
  summarise = TRUE,
  n.iter = 1000,
  plot_diagnostics = TRUE
)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}

\item{summarise}{logical TRUE or FALSE. Plot the realisations as a summarised
"ribbon" showing 50% and 95% intervals (faster), or as a spaghetti plot
showing individual realisations. Defaults to TRUE (ribbon).}

\item{n.iter}{The number of iterations of the model to plot, defaults to
1000.}

\item{plot_diagnostics}{logical, include diagnostic plots: traceplot of
log-posterior, hierarchical accumulations rates, memory parameter. Defaults
to TRUE.}
}
\value{
A ggplot2 object
}
\description{
Plots the HAMStR modelled age ~ depth relationship together with
  the depths, ages, and age uncertainties in the observed data. A random
  sample of size \code{n.iter} of the iterations of the posterior
  distribution are plotted as grey lines. The observed data are plotted as
  points with +- 2*se error bars.
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error,
  K = c(10, 10), nu = 6,
  acc_mean_prior = 20,
  mem_mean = 0.5, mem_strength= 10,
  inflate_errors = 0,
  iter = 2000, chains = 3)

# With age models summarised as a ribbon. Faster than spaghetti plots.
plot_hamstr(fit)

# With age models as spaghetti plots. Can see individual realisations, but slower to plot.
plot_hamstr(fit, summarise = FALSE)
}
}
