% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_hamstr_acc_rates}
\alias{plot_hamstr_acc_rates}
\title{Plot accumulation rates}
\usage{
plot_hamstr_acc_rates(
  hamstr_fit,
  units = c("depth_per_time", "time_per_depth")
)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{

}
\description{
Plot accumulation rates
}
\keyword{internal}
