% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_bacon.R
\name{plot_hierarchical_acc_rate}
\alias{plot_hierarchical_acc_rate}
\title{Plot the hierarchical accumulation rate parameters}
\usage{
plot_hierarchical_acc_rate(adam_fit)
}
\arguments{
\item{adam_fit}{}
}
\value{

}
\description{
Plot the hierarchical accumulation rate parameters
}
