% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_memory_prior_posterior}
\alias{plot_memory_prior_posterior}
\title{Plot Memory Prior and Posterior}
\usage{
plot_memory_prior_posterior(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{
A ggplot2 object
}
\description{
Plot Memory Prior and Posterior
}
\keyword{0,}
\keyword{0.5,}
\keyword{10),}
\keyword{10,}
\keyword{20,}
\keyword{2000,}
\keyword{3)}
\keyword{6,}
\keyword{<-}
\keyword{=}
\keyword{K}
\keyword{MSB2K$age,}
\keyword{MSB2K$depth,}
\keyword{MSB2K$error,}
\keyword{\dontrun{}
\keyword{acc_mean_prior}
\keyword{c(10,}
\keyword{chains}
\keyword{depth}
\keyword{fit}
\keyword{hamstr(}
\keyword{inflate_errors}
\keyword{internal}
\keyword{iter}
\keyword{mem_mean}
\keyword{mem_strength}
\keyword{nu}
\keyword{obs_age}
\keyword{obs_err}
\keyword{plot_memory_prior_posterior(fit)}
\keyword{}}
