% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_prior_posterior_hist}
\alias{plot_prior_posterior_hist}
\title{Plot a Prior and Posterior}
\usage{
plot_prior_posterior_hist(prior, posterior)
}
\arguments{
\item{prior}{}

\item{posterior}{}
}
\value{
A ggplot2 object
}
\description{
Plot a Prior and Posterior
}
\keyword{internal}
