% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_hamstr.R
\name{plot_summary_age_models}
\alias{plot_summary_age_models}
\title{Plot Summary of Posterior Age Models}
\usage{
plot_summary_age_models(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{
A ggplot2 object
}
\description{
Plot Summary of Posterior Age Models
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error,
  K = c(10, 10), nu = 6,
  acc_mean_prior = 20,
  mem_mean = 0.5, mem_strength = 10,
  inflate_errors = 0,
  iter = 2000, chains = 3)

plot_summary_age_models(fit)
}
}
\keyword{internal}
