% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{predict.hamstr_fit}
\alias{predict.hamstr_fit}
\title{Interpolate Age Models at Given Depths}
\usage{
\method{predict}{hamstr_fit}(object, depth = NULL)
}
\arguments{
\item{object}{}

\item{depth}{a vector of depths at which to interpolate the age models. 
If left NULL, the depths of the age control points are used.}
}
\value{

}
\description{
Method for generic function predict. Returns the posterior age
models interpolated to new depths given in depth.
}
