% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{predict.hamstr_fit}
\alias{predict.hamstr_fit}
\title{Interpolate Age Models at Given Depths}
\usage{
\method{predict}{hamstr_fit}(
  object,
  type = c("age_models", "acc_rates"),
  depth = c("modelled", "data"),
  tau = 0,
  kern = c("U", "G", "BH"),
  ...
)
}
\arguments{
\item{object}{hamstr_fit object}

\item{depth}{Defaults to "modelled", which returns the modelled depths.
"data" returns age models at the depths of the observations, or a numerical
 vector to specify depths. Accumulation rates are only returned at the 
 modelled depths.}
}
\value{

}
\description{
Method for generic function predict. Returns the posterior age
models interpolated to new depths given in depth.
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error
  )

predict(fit, depth = seq(1, 100, by = 10))
predict(fit, depth = "data")
predict(fit)
predict(fit, type = "acc_rates")
}

}
