% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sarnie.R
\name{sarnie}
\alias{sarnie}
\title{sarnie}
\usage{
sarnie(
  depth,
  obs_age,
  obs_err,
  K1 = 10,
  K = 100,
  nu = 6,
  record_prior_acc_mean_mean = 20,
  record_prior_acc_mean_shape = 1.5,
  record_prior_acc_shape_mean = 1.5,
  record_prior_acc_shape_shape = 1.5,
  mem_mean = 0.7,
  mem_strength = 4,
  inflate_errors = 0,
  iter = 2000,
  chains = 3,
  ...
)
}
\arguments{
\item{depth}{Vector of depths from an observed age-depth profile}

\item{obs_age}{Observed age at each depth}

\item{obs_err}{Error associated with each observed age (1 standard error)}

\item{K1}{Number of coarse level sections}

\item{K}{Number of sections into which the profile will be divided, must be a multiple of K1}

\item{nu}{Degrees of freedom for the Student-t distributed error model.
Defaults to 6, which is equivalent to the default parameterisation of
t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian
error model, (nu = 100 should do it).}

\item{record_prior_acc_mean_mean}{hyperprior parameter for the prior on the overall mean acc.rate for 
the record, set to the mean acc.rate for the record from a linear model or similar}

\item{record_prior_acc_mean_shape}{shape for the prior on the overall mean, leave at 1.5 for now 
acc.rate for the record}

\item{record_prior_acc_shape_mean}{hyperprior parameters for the prior on 
the shape of the innovations distribution, leave at 1.5 for now}

\item{record_prior_acc_shape_shape}{hyperprior parameters for the prior on 
the shape of the innovations distribution, leave at 1.5 for now}

\item{mem_mean}{Hyper-parameter: a parameter of the Beta prior distribution
on "memory", i.e. the autocorrelation parameter in the underlying AR1
model. The prior on the correlation between layers is scaled according to
the thickness of the layers *delta_c*, which is determined by the total
length of the profile and the parameter *K*. mem_mean sets the mean value
for *R* (defaults to 0.7), while *w* = R^(delta_c)}

\item{mem_strength}{Hyper-parameter: sets the strength of the memory prior,
defaults to 4 as in Bacon 2.2}

\item{inflate_errors}{logical, 1 or 0. If set to 1, observation errors are 
inflated so that data are consistent with a Bacon type monotonic age-depth model}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}

\item{chains}{A positive integer specifying the number of Markov chains. 
    The default is 4.}

\item{...}{Additional arguments can be \code{chain_id}, \code{init_r}, 
    \code{test_grad}, \code{append_samples}, \code{refresh},
    \code{enable_random_init}. See the documentation in \code{\link[rstan]{stan}}.}

\item{acc_mean}{The mean sediment accumulation rate for the Gamma prior on
sedimentation rate innovations}

\item{acc_alpha}{Sets the alpha (shape) parameter for the Gamma prior on
sedimentation rate innovations. The variance of the innovations acc_var =
acc_mean^2^ / acc_alpha. The default is acc_alpha = 1.5 which is equivalent
to the default shape = 1.5 parameter in Bacon 2.2}
}
\value{
Returns a list composed of the output from the Stan sampler .$fit,
  and the list of data passed to the sampler
}
\description{
sarnie
}
\examples{

fit <- sarnie(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error,
  K = 100, K1 = 10, nu = 6,
  record_prior_acc_mean_mean = 20,
  mem_mean = 0.7, mem_strength = 4,
  inflate_errors = 0,
  iter = 2000, chains = 3)

print(fit$fit, par = c("record_acc_mean"))

plot_stan_bacon(fit, 100, plot_priors = FALSE)

}
