% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sarnie.R
\name{sarnie}
\alias{sarnie}
\title{sarnie}
\usage{
sarnie(
  depth,
  obs_age,
  obs_err,
  K1 = 10,
  K = 100,
  nu = 6,
  top_depth = NULL,
  bottom_depth = NULL,
  pad_top_bottom = FALSE,
  record_prior_acc_mean_mean = 20,
  record_prior_acc_mean_shape = 1.5,
  record_prior_acc_shape_mean = 1.5,
  record_prior_acc_shape_shape = 1.5,
  section_acc_shape = 1.5,
  mem_mean = 0.7,
  mem_strength = 4,
  inflate_errors = 0,
  iter = 2000,
  chains = 3,
  ...
)
}
\arguments{
\item{depth}{Vector of depths from an observed age-depth profile}

\item{obs_age}{Observed age at each depth}

\item{obs_err}{Error associated with each observed age (1 standard error)}

\item{K1}{Number of coarse level sections}

\item{K}{Number of fine level sections into which the profile will be divided}

\item{nu}{Degrees of freedom for the Student-t distributed error model.
Defaults to 6, which is equivalent to the default parameterisation of
t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian
error model, (nu = 100 should do it).}

\item{top_depth, bottom_depth}{the top and bottom depths of the age-depth
model. Must encompass the range of the data. Default to the shallowest and
deepest data points unless \code{pad_top_bottom = TRUE}}

\item{pad_top_bottom}{logical, pad the length of the age-depth model by 5% on
each end}

\item{record_prior_acc_mean_mean, record_prior_acc_mean_shape}{hyperparameters
for the hyperprior on the overall mean accumulation rate for the record.
Units are depth / obs_age. E.g. if depth is in cm and age in kyr then the
accumulation rate is in cm/kyr. ...mean_shape sets the shape of the
hyperprior for the mean. Defaults to 1.5.}

\item{record_prior_acc_shape_mean, record_prior_acc_shape_shape}{hyperparameters for the hyperprior on the shape of the distribution of K1
coarse level accumulation rates. Defaults to 1.5, 1.5.}

\item{section_acc_shape}{hyperparameter for the prior on the shape of the K
fine level accumulation rates. Defaults to 1.5 - as for Bacon 2.2.}

\item{mem_mean}{Hyperparameter: a parameter of the Beta prior distribution on
"memory", i.e. the autocorrelation parameter in the underlying AR1 model.
The prior on the correlation between layers is scaled according to the
thickness of the layers *delta_c*, which is determined by the total length
of the profile and the parameter *K*. mem_mean sets the mean value for *R*
(defaults to 0.7), while *w* = R^(delta_c)}

\item{mem_strength}{Hyperparameter: sets the strength of the memory prior,
defaults to 4 as in Bacon 2.2}

\item{inflate_errors}{logical. If set to TRUE, observation errors are
inflated so that data are consistent with a Bacon type monotonic age-depth
model}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}

\item{chains}{A positive integer specifying the number of Markov chains. 
    The default is 4.}

\item{...}{Additional arguments can be \code{chain_id}, \code{init_r}, 
    \code{test_grad}, \code{append_samples}, \code{refresh},
    \code{enable_random_init}. See the documentation in \code{\link[rstan]{stan}}.}
}
\value{
Returns a list composed of the output from the Stan sampler .$fit,
  and the list of data passed to the sampler
}
\description{
sarnie
}
\examples{

dontrun{

fit <- sarnie(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error,
  K = 100, K1 = 10, nu = 6,
  record_prior_acc_mean_mean = 20,
  mem_mean = 0.7, mem_strength = 4,
  inflate_errors = 0,
  iter = 2000, chains = 3)

print(fit$fit, par = c("record_acc_mean"))

plot_stan_bacon(fit, 100, plot_priors = TRUE)

}
}
