% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_bacon.R
\name{stan_bacon}
\alias{stan_bacon}
\title{stan_bacon}
\usage{
stan_bacon(
  depth,
  obs_age,
  obs_err,
  hiatus_depth = NULL,
  hiatus_length = NULL,
  hiatus_shape = 1,
  hiatus_interval = 0.1,
  K = 10,
  nu = 6,
  acc_mean,
  acc_alpha = 1.5,
  mem_mean = 0.7,
  mem_strength = 4,
  iter = 2000,
  chains = 3
)
}
\arguments{
\item{depth}{Vector of depths from an observed age-depth profile}

\item{obs_age}{Observed age at each depth}

\item{obs_err}{Error associated with each observed age (1 standard error)}

\item{hiatus_depth}{Depths of suspected hiatuses, as vector}

\item{hiatus_length}{Priors for lengths of suspected hiatuses (in time units)}

\item{hiatus_shape}{Shape parameter for the gamma prior on hiatus length,
defaults to 1 which results in a gamma distribution with 1 degree of
freedom, i.e. an exponential distribution with mean and sd equal to
hiatus_length}

\item{hiatus_interval}{Depth interval over which hiatus occurs. Should be
small but cannot be zero}

\item{K}{Number of sections into which the profile will be divided}

\item{nu}{Degrees of freedom for the Student-t distributed error model.
Defaults to 6, which is equivalent to the default parameterisation of
t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian
error model, (nu = 100 should do it).}

\item{acc_mean}{The mean sediment accumulation rate for the Gamma prior on
sedimentation rate innovations}

\item{acc_alpha}{Sets the alpha (shape) parameter for the Gamma prior on
sedimentation rate innovations. The variance of the innovations acc_var =
acc_mean^2^ / acc_alpha. The default is acc_alpha = 1.5 which is equivalent
to the default shape = 1.5 parameter in Bacon 2.2}

\item{mem_mean}{Hyper-parameter: a parameter of the Beta prior distribution
on "memory", i.e. the autocorrelation parameter in the underlying AR1
model. The prior on the correlation between layers is scaled according to
the thickness of the layers *delta_c*, which is determined by the total
length of the profile and the parameter *K*. mem_mean sets the mean value
for *R* (defaults to 0.7), while *w* = R^(delta_c)}

\item{mem_strength}{Hyper-parameter: sets the strength of the memory prior,
defaults to 4 as in Bacon 2.2}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}

\item{chains}{A positive integer specifying the number of Markov chains. 
    The default is 4.}
}
\value{
Returns a list composed of the output from the Stan sampler .$fit,
  and the list of data passed to the sampler
}
\description{
stan_bacon
}
\examples{

dontrun{
K_for_5cm <- round(diff(range(MSB2K$depth)) / 5)
fit <- stan_bacon(
  depth = MSB2K$depth,
  obs_age = MSB2K$age.cal,
  obs_err = MSB2K$age.cal.sd,
  K = K_for_5cm, nu = 6,
  acc_mean = 20, acc_alpha = 1.5,
  mem_mean = 0.7, mem_strength = 4,
  iter = 2000, chains = 3)

  print(fit$fit, par = c("R", "w", "c_ages"))

  plot_stan_bacon(fit, 1000)
}
}
