% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_age_models}
\alias{summarise_age_models}
\title{Summarise Posterior Age Models}
\usage{
summarise_age_models(hamstr_fit, probs)
}
\arguments{
\item{hamstr_fit}{The object returned from \code{stan_hamstr}.}
}
\value{
data.frame / tibble
}
\description{
Extracts the summary statistics of posterior age models and attaches the depths
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error)

summarise_age_models(fit)
}
}
\keyword{internal}
