% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_age_models}
\alias{summarise_age_models}
\title{Summarise Posterior Age Models}
\usage{
summarise_age_models(adam_fit)
}
\arguments{
\item{adam_fit}{}
}
\value{

}
\description{
Extracts the summary statistics of posterior age modells and attached the depths
}
