% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_age_models}
\alias{summarise_age_models}
\title{Summarise Posterior Age Models}
\usage{
summarise_age_models(hamstr_fit)
}
\arguments{
\item{hamstr_fit}{an hamstr_fit object or hamstr_interpolated_ages object}
}
\value{

}
\description{
Extracts the summary statistics of posterior age models and attached the depths
}
