% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_hamstr_acc_rates}
\alias{summarise_hamstr_acc_rates}
\title{Summarise accumulation rates}
\usage{
summarise_hamstr_acc_rates(
  hamstr_fit,
  tau = 0,
  kern = c("U", "G", "BH"),
  probs = c(0.025, 0.159, 0.25, 0.5, 0.75, 0.841, 0.975)
)
}
\arguments{
\item{hamstr_fit}{a hamstr_fit object}

\item{tau}{scale of the smoothing kernel in depth units. If tau > 0,
accumulation rates are smoothed (filtered) before summary statistics are
calculated, so that the accumulation rate at a given depth corresponds to the
average rate over the depth interval tau. Default to 0.}

\item{kern}{choice of smoothin kernal. U for uniform (moving average), G for
Gaussian, BH for Berger and Heath (exponential). Defaults to U}
}
\value{
a tibble containing the summarised posterior accumulation rates
}
\description{
Summarise accumulation rates
}
\keyword{internal}
