% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_hamstr_parameters}
\alias{summarise_hamstr_parameters}
\title{Summarise Hamstr Parameters}
\usage{
summarise_hamstr_parameters(
  object,
  pars = c("alpha[1]", "R", "w", "L", "D", "H_depth", "H_length"),
  probs = c(0.025, 0.159, 0.25, 0.5, 0.75, 0.841, 0.975)
)
}
\arguments{
\item{object}{A hamstr_fit object}

\item{pars}{Character vector of parameters to include}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}
}
\value{
a tibble of summarised posterior of hamstr parameters
}
\description{
Summarise Hamstr Parameters
}
\keyword{internal}
