% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_new_ages}
\alias{summarise_new_ages}
\title{Summarise Interpolated Posterior Age Models}
\usage{
summarise_new_ages(new_ages)
}
\arguments{
\item{new_ages}{}
}
\value{

}
\description{
Summarise Interpolated Posterior Age Models
}
