% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summarise_q}
\alias{summarise_q}
\title{Summarise to Quantiles and Moments}
\usage{
summarise_q(dat, var, probs = c(0.025, 0.25, 0.5, 0.75, 0.975))
}
\arguments{
\item{dat}{A dataframe or tibble}

\item{var}{The variable to summarise (unquoted)}

\item{probs}{The quantiles at which to summarise}
}
\value{

}
\description{
Summarise to Quantiles and Moments
}
\keyword{internal}
