% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summary.hamstr_fit}
\alias{summary.hamstr_fit}
\title{Summarise hamstr models}
\usage{
\method{summary}{hamstr_fit}(object, type = c("age_models", "acc_rates", "pars"), ...)
}
\arguments{
\item{object}{hamstr_fit object}

\item{type}{age models "age_models" or accumulation rates "acc_rates"}

\item{...}{additional arguments to hamstr summary methods}
}
\value{

}
\description{
Summarise hamstr models
}
\examples{
\dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error,
  K = c(10, 10))

summary(fit)
summary(fit, type = c("acc_rates"))
}
}
