% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summary.hamstr_interpolated_ages}
\alias{summary.hamstr_interpolated_ages}
\title{Summarise Age Models That Have Been Interpolated}
\usage{
\method{summary}{hamstr_interpolated_ages}(object, ...)
}
\arguments{
\item{object}{hamstr_interpolated_ages object}

\item{...}{additional arguments to hamstr summary methods}
}
\value{
a tibble containing the summarised posterior
}
\description{
summary method for class "hamstr_interpolated_ages"
}
